+++ anchore_engine/services/apiext/api/controllers/system.py
@@ -1,7 +1,5 @@
 from connexion import request
-import copy
 import json
-import datetime
 
 # anchore modules
 import anchore_engine.apis
@@ -13,7 +11,6 @@ from anchore_engine.clients.services.catalog import CatalogClient
 from anchore_engine.clients.services.policy_engine import PolicyEngineClient
 import anchore_engine.common
 import anchore_engine.subsys.servicestatus
-from anchore_engine.db.entities.catalog import QueueItem
 import anchore_engine.configuration.localconfig
 from anchore_engine.configuration.localconfig import GLOBAL_RESOURCE_DOMAIN
 from anchore_engine.apis.context import ApiRequestContextProxy
@@ -378,16 +375,3 @@ def describe_error_codes():
 
     return return_object, httpcode
 
-@authorizer.requires([])
-def get_webhook_schema():
-    request_inputs = anchore_engine.apis.do_request_prep(request, default_params={})
-    return_object = {}
-    httpcode = 500
-    try:
-        return_object = QueueItem.to_schema()
-        httpcode = 200
-    except Exception as err:
-        return_object = anchore_engine.common.helpers.make_response_error(err, in_httpcode=httpcode)
-        httpcode = return_object['httpcode']
-
-    return return_object, httpcode
