+++ anchore_engine/services/apiext/swagger/swagger.yaml
@@ -1642,18 +1642,6 @@ paths:
             items:
               $ref: "#/definitions/GateSpec"
       x-swagger-router-controller: anchore_engine.services.apiext.api.controllers.system
-  /webhooks/schema:
-    get:
-      summary: Get the JSON Schema of the webhooks' payload generated by the notification service
-      description: Prints out a JSON schema representing the JSON payload a webhook would send for a notification
-      operationId: get_webhook_schema
-      x-anchore-authz-action: None
-      tags:
-        - Webhooks
-      responses:
-        200:
-          description: the JSON Schema
-      x-swagger-router-controller: anchore_engine.services.apiext.api.controllers.system
   /system/error_codes:
     get:
       summary: Describe anchore engine error codes.
@@ -2599,6 +2587,18 @@ paths:
           description: Internal error
           schema:
             $ref: "#/definitions/ApiErrorResponse"
+  /webhooks/schema:
+    get:
+      summary: Get the JSON Schema of the webhooks' payload generated by the notification service
+      description: Prints out a JSON schema representing the JSON payload a webhook would send for a notification
+      operationId: get_webhook_schema
+      x-anchore-authz-action: None
+      tags:
+        - Webhooks
+      responses:
+        200:
+          description: the JSON Schema
+      x-swagger-router-controller: anchore_engine.services.apiext.api.controllers.webhooks
 definitions:
   PaginationProperties:
     type: object
